+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix "Publication type" column is empty in "Titles" accordion of "Package" record. (UIEH-1487)
 * Populate `resourceName` when adding an eHoldings resource to an agreement line via "eHoldings". (ERM-4007)
 * Update Tags label in "Titles" accordion of "Package" record. (UIEH-1488)
+* Show `createdDate` when `updatedDate` is not available in Access Status Types metadata. (UIEH-1490)
 
 ## [11.1.1] (https://github.com/folio-org/ui-eholdings/tree/v11.1.1) (2026-05-28)